+++ app/src/main/java/com/stoyan/weatherful/view_utils/recyclerview_utils/forecast_recyclerview/ForecastViewHolder.java
@@ -73,7 +73,7 @@ public class ForecastViewHolder extends RecyclerView.ViewHolder {
     }
 
     private void setRainChance(final Data data) {
-        double probability = Double.parseDouble(data.getPrecipProbability()) * 100;
+        Float probability = Float.parseFloat(data.getPrecipProbability()) * 100;
         String displayRainChance = "Chance of raining: " + probability + "%";
         tvRainChance.setText(displayRainChance);
     }
